backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
neo4j==5.21.0
pydantic==2.7.4
python-dotenv==1.0.1
